Use void* + element size to handle any type
(not tested yet, but compiles)
Using void * + storing the size of an element, the same struct is able to store
any type of any size.
Sadly this means that type safety is lost, and it also is impossible to insert rvalues
and return a popped element.
sorting / comarisons functions are noot implemented yet.
Sorting would require a swap function on a value of arbitrary size (maybe use a static vector as temp storage ?)
Sorting, count_elems and is_sorted require a compparison functions, whic s manageable